requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
model-bakery==1.19.5
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytest==8.3.2
pytest-django==4.8.0
pytest-testdox==3.1.0
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.1
